+++ src/components/SearchByIDA/index.tsx
@@ -13,16 +13,18 @@ import ToggleSwitch from 'components/ToggleSwitch';
 
 import cssBinder from 'styles/cssBinder';
 
+import blocks from 'styles/blocks.css';
 import local from './style.css';
+import searchPageCss from 'pages/Search/style.css';
 import search from 'components/IPScan/Search/style.css';
 import Button from 'components/SimpleCommonComponents/Button';
+
+const css = cssBinder(local, blocks, search, searchPageCss);
 
 const SchemaOrgData = loadable({
   loader: () => import(/* webpackChunkName: "schemaOrg" */ 'schema_org'),
   loading: () => null,
 });
-
-const css = cssBinder(local, search);
 
 type Props = {
   customLocation: InterProLocation;
@@ -143,8 +145,11 @@ export class SearchByIDA extends PureComponent<Props, State> {
     if (ignoreFromURL !== undefined && ignoreFromURL.trim() === '')
       ignore.push('');
     return (
-      <div className={css('vf-stack', 'vf-stack--400')}>
-        <div className={css('ida-search')}>
+      <section className={css('vf-stack', 'vf-stack--400')}>
+        <div className={css('simple-box')}>
+          <header>
+            Search for proteins with a specific domain architecture
+          </header>
           <SchemaOrgData
             data={{
               name: 'Search By Domain Architecture IDA',
@@ -154,9 +159,6 @@ export class SearchByIDA extends PureComponent<Props, State> {
           />
           <div className={css('vf-stack', 'vf-stack--400')}>
             <div className={css('search-input')}>
-              <h3 className={css('light')}>
-                Search for proteins with a specific domain architecture
-              </h3>
               <div className={css('description')}>
                 <p>
                   Domain architectures are derived from matches to Pfam models.
@@ -280,7 +282,7 @@ export class SearchByIDA extends PureComponent<Props, State> {
             </div>
           </div>
         </div>
-      </div>
+      </section>
     );
   }
 }
